ingest/requirements.txt
langchain>=0.3.0
langchain-community>=0.3.0
sentence-transformers>=2.2.0
supabase>=2.0.0
python-dotenv>=1.0.0
pyyaml>=6.0
gitpython>=3.1.0
